refactor( src/app/config/themes/components.ts src/app/config/themes/index.tsx src/app/config/themes/palette.ts src/app/config/themes/types.ts src/app/config/themes/typography.ts) criamos uma pasta theme com novos arquivos, cada um para uma funcao antes exercida pelo unico arquivo theme e alteramos todas as importações @stephanybrazeir0 @Geoziihdev @Lip29bm
