+++ src/app/config/themes/typography.ts
@@ -1,0 +1,74 @@
+export const typography = {
+  h1: {
+    fontWeight: 500,
+    letterSpacing: 0.15,
+    lineHeight: "normal",
+    "@media (min-width:0px)": {
+      fontSize: "1.5rem"
+    },
+    "@media (min-width:600px)": {
+      fontSize: "2rem"
+    },
+    "@media (min-width:900px)": {
+      fontSize: "2.5rem"
+    },
+    "@media (min-width:1200px)": {
+      fontSize: "3rem"
+    },
+    "@media (min-width:1536px)": {
+      fontSize: "3.5rem"
+    }
+  },
+  h2: {
+    lineHeight: 1.6,
+    fontWeight: 500,
+    letterSpacing: 0.15,
+    "@media (min-width:0px)": {
+      fontSize: "1.5rem"
+    },
+    "@media (min-width:600px)": {
+      fontSize: "1.8rem"
+    },
+    "@media (min-width:900px)": {
+      fontSize: "2.2rem"
+    },
+    "@media (min-width:1200px)": {
+      fontSize: "2.4rem"
+    },
+    "@media (min-width:1536px)": {
+      fontSize: "2.4rem"
+    }
+  },
+  h3: {
+    lineHeight: 1.6,
+    fontWeight: 500,
+    letterSpacing: 0.15,
+    "@media (min-width:0px)": {
+      fontSize: "1.2rem"
+    },
+    "@media (min-width:600px)": {
+      fontSize: "1.4rem"
+    },
+    "@media (min-width:900px)": {
+      fontSize: "1.5rem"
+    },
+    "@media (min-width:1200px)": {
+      fontSize: "1.6rem"
+    },
+    "@media (min-width:1536px)": {
+      fontSize: "1.6rem"
+    }
+  },
+  body1: {
+    fontSize: "1rem",
+    lineHeight: 1.5,
+    fontWeight: 400,
+    letterSpacing: 0.15
+  },
+  caption: {
+    fontSize: "1.4rem",
+    lineHeight: 1.6,
+    fontWeight: 500,
+    letterSpacing: 0.15
+  }
+}
