+++ src/components/ClickButton/index.tsx
@@ -2,11 +2,11 @@ import * as React from "react"
 import { styled } from "@mui/material/styles"
 import Button, { ButtonProps } from "@mui/material/Button"
 import Stack from "@mui/material/Stack"
-import { theme } from "@/app/config/theme"
+import { theme } from "@/app/config/themes"
 import { ReactNode } from "react"
 
 
-const ColorButton = styled(Button)<ButtonProps>(() => ({    
+const ColorButton = styled(Button)<ButtonProps>(() => ({
     "&:hover": {
         backgroundColor: theme.palette.buttonHover?.main,
         color: theme.palette.buttonHover?.contrastText,
@@ -22,7 +22,7 @@ type CardProps = {
 }
 
 export const ClickButton = ({ title, click, endIcon, backIcon }: CardProps) =>
-    <aside>        
+    <aside>
         <Stack spacing={2} direction="row" onClick={click}>
             <ColorButton sx={theme.customStyles.button} variant="contained">{backIcon}{title}{endIcon}</ColorButton>
         </Stack>
