+++ src/components/Description/DescriptionReference/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Box, Grid, Link, Typography } from "@mui/material";
-import { theme } from "@/app/config/theme";
+import { Grid, Link } from "@mui/material";
+import { theme } from "@/app/config/themes";
 import ReactMarkdown from "react-markdown";
 
 interface DescriptionReferenceProps {
@@ -26,9 +26,9 @@ export const DescriptionReference: React.FC<DescriptionReferenceProps> = ({
     ),
   };
   return (
-    <Grid container alignItems="stretch" sx={{...theme.customStyles.description}}>
+    <Grid container alignItems="stretch" sx={{ ...theme.customStyles.description }}>
       <Grid item xl={12} lg={12} md={12} sm={12} xs={12}>
-      <ReactMarkdown components={components}>{text}</ReactMarkdown>
+        <ReactMarkdown components={components}>{text}</ReactMarkdown>
       </Grid>
     </Grid>
   );
