+++ src/components/LoginButton/index.tsx
@@ -2,10 +2,10 @@ import * as React from "react"
 import { styled } from "@mui/material/styles"
 import Button, { ButtonProps } from "@mui/material/Button"
 import Stack from "@mui/material/Stack"
-import { theme } from "@/app/config/theme"
+import { theme } from "@/app/config/themes"
 
 
-const ColorButton = styled(Button)<ButtonProps>(() => ({    
+const ColorButton = styled(Button)<ButtonProps>(() => ({
     "&:hover": {
         backgroundColor: theme.palette.buttonHover?.main,
         color: theme.palette.buttonHover?.contrastText,
@@ -18,8 +18,8 @@ type CardProps = {
 }
 
 export const LoginButton = ({ click }: CardProps) =>
-    <aside>        
+    <aside>
         <Stack spacing={2} direction="row" onClick={click}>
-            <ColorButton sx={{... theme.customStyles.button , padding: "5px 16px"}} variant="contained">LOGIN</ColorButton>
+            <ColorButton sx={{ ...theme.customStyles.button, padding: "5px 16px" }} variant="contained">LOGIN</ColorButton>
         </Stack>
     </aside>
