+++ src/components/NavBar/index.tsx
@@ -3,7 +3,7 @@ import AppBar from "@mui/material/AppBar"
 import Toolbar from "@mui/material/Toolbar"
 import Container from "@mui/material/Container"
 import { Box } from "@mui/material"
-import { theme } from "../../app/config/theme"
+import { theme } from "@/app/config/themes"
 import { MobileMenu } from "./MobileNavBar"
 import { WebMenu } from "./WebNavBar"
 import { Session } from "next-auth"
@@ -18,7 +18,7 @@ const ResponsiveAppBar: React.FC<NavBarProps> = ({ list, session }) => {
   return (
     <AppBar
       position="fixed"
-      sx={{ backgroundColor: theme.palette.bgColor?.main }}
+      sx={{ backgroundColor: theme.palette.bgColor?.main}}
     >
       <Box component="nav">
         <Container maxWidth="xl">
